Increase wdspec long timeout to three minutes.
The long timeout for WPT wdspec tests is currently two minutes,
but with modifications to make more session tests run in Firefox,
it has been discovered that the long timeout is not sufficient time
for Firefox to consistently run the New Session command tests in
automation without timing out.

This increases the long timeout to three minutes (180 seconds).

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1470646
gecko-commit: da9689d7801b57bc79b6a82fa1b7678d44bed170
gecko-integration-branch: autoland
gecko-reviewers: whimboo