ImageLoading: Don't reschedule a microtask on attach if one is pending.
This patch ensures that we don't reschedule an image loading microtask
if there is one pending already. This is important to ensure that image
decode microtasks see the image data (ie the image loading microtask
runs first) if there are no other changes only insertion into a
document.

The existing tests that were added for this still pass (from crrev.com/cdd156337ada7)
  http/tests/misc/image-blocked-src-change.html
  http/tests/misc/image-blocked-src-no-change.html
  http/tests/xmlhttprequest/xmlhttprequest-image-not-loaded-svg.svg
  http/tests/xmlhttprequest/xmlhttprequest-image-not-loaded.html

The new test fails without the change and passes with.

Bug: 864306
Change-Id: I1d3d2ed4c12b53fab931e55a85c4f5cd18fe1cc4
R=chrishtr@chromium.org